refactor(main): extract the watch-root registry into its own module
file-watcher.js reached 1335 lines when #227 merged the registry into it. The
registry is pure state — a plan of watch roots, their per-root transitions, and
W derived over the whole plan — while everything around it spawns powershell,
talks to chokidar and writes sensor-health records. Split, the state machine is
readable on its own: file-watcher.js 1335 -> 1058, watch-root-registry.js 301.

src/main/watch-root-registry.js requires nothing but sensor-health, and only for
the state enum it names W with. No I/O: no chokidar, no fs, no timers. It stores
the FSWatcher objects it is given and never creates, closes or inspects one.

Two moved functions changed beyond import wiring, both so the fs-chokidar record
stays owned by file-watcher.js:
- markRootReady / markRootErrored return whether the plan moved instead of
  calling applyWatchPlaneHealth themselves. bindWatcherEvents writes the record
  on true, reproducing both old early returns — an unconditional write would
  re-derive W over an empty plan, which throws by design.
- markRootRegistrationFailed / markRootErrored take the already-normalized
  message. healthErrorMessage stays in file-watcher.js with its four other
  callers, matching the per-module copy in process-scanner and network-monitor.
hasPlannedRoot is new: setupFileWatchers read the plan Map directly to skip a
group a preflight proved absent.

attribution.js: the AttributionEvidence union in events.ts now carries
population-unavailable, so the JSDoc no longer claims it is missing, and it
names what actually holds the two halves together (a test, not tsc).

No behaviour change: the six file-watcher suites and the full suite (1794 pass,
4 skip) are green with no assertion touched. Verified from the new home by
deleting the ERRORED guard in markRootReady and watching the terminal-state
test go red, then reverting.
